+++ docs/blog3.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html lang="en">
 	<head>
-    <title>Learning CSS</title>
+    <title>{{title}}</title>
   		<meta charset="utf-8">
 		<meta name="viewport" content="width=device-width, initial-scale=1">
 		<link href="https://fonts.googleapis.com/css?family=PT+Sans&display=swap" rel="stylesheet"> 
@@ -42,7 +42,7 @@
 					<div class="posts">
 						<div>
            						<h1>Kickstart coding, Week 3</h1>
-						<p>11/11/19</p>
+						<p>11/08/19</p>
             <p>This week we started learning Python</p><p>I have played around with several python tutuorials but have never really built projects.</p>k
             <p>I am really enjoying the project based learning and look forward to getting more comfortable with it!</p>
